Updated functions

+++ SET_Solver.py
@@ -7,17 +7,8 @@ from SET_Classifier import MultiClassCNN
 
 # global variables
 condition = False
-
-
-# number = {1, 2, 3}
-#          [0, 1, 2]
-# color = {red, purple,  green}
-#         [0  ,     1,      2]
-# shape = {squiggle, diamond, oval}
-#         [0       ,       1,    2]
-# fill = {full, half, empty}
-#        [0   ,    1,    2]
-# visit https://www.setgame.com/set/puzzle to scrape clear images of all 81 cards
+verbose = True
+reset = True
 
 def split(img):
     """
@@ -43,19 +34,26 @@ def classify():
     Classifies each image in Test_Images and returns the results
     :return output: list of lists of card features
     """
-    # labels csv contains the conversion from class name to card features
-    labels = pd.read_csv("labels.csv", index_col=0)
-    classifier_model = torch.load("model.pt")
+    number_model = torch.load("model0.pt")
+    color_model = torch.load("model1.pt")
+    shape_model = torch.load("model2.pt")
+    fill_model = torch.load("model3.pt")
     output = []
     for i in range(12):
         img = torchvision.io.read_image("Test_Images/" + str(i) + ".png").unsqueeze(dim=0).float()
         if img.shape[1] == 3:
             img = torch.cat((img, torch.zeros(1, 1, 167, 258)), dim=1)
-        class_ret = classifier_model(img)
-        result = torch.argmax(class_ret, dim=1).item()
-        # print("Image: ", i, "  Result: ", result)
-        ret = labels.loc[result].to_numpy()
-        output.append(ret)
+        
+        number = number_model(img).argmax(dim=1).item()
+        color = color_model(img).argmax(dim=1).item()     
+        shape = shape_model(img).argmax(dim=1).item()
+        fill = fill_model(img).argmax(dim=1).item()        
+        label = [number, color, shape, fill]
+        if verbose:
+            if i == 0:
+                print("Classifying: ")
+            print("Image: ", i, "  Result: ", label)
+        output.append(label)
     return output
 
 
@@ -65,11 +63,15 @@ def find(results):
     :param results: a list of lists of card features
     :return sets: list of sets
     """
+    if verbose:
+        print("Finding sets: ")
     sets = []
     for i in range(len(results)):
         for j in range(i + 1, len(results)):
             for k in range(j + 1, len(results)):
                 if is_set(results[i], results[j], results[k]):
+                    if verbose:
+                        print("Set found: |", results[i], results[j], results[k], "|")
                     sets.append([i, j, k])
     return sets
 
@@ -102,32 +104,37 @@ def show(output, image_name):
     color_dict = {0: "red", 1: "green", 2: "purple", 3: "yellow",
                   4: "blue", 5: "orange", 6: "pink", 7: "black",
                   8: "white", 9: "grey", 10: "brown", 11: "cyan"}
-    # go back to building the image from composites
+    # building the image from composites
     img = Image.new("RGB", (258 * 4, 167 * 3))
     for i in range(12):
         img_i = Image.open("Test_Images/" + str(i) + ".png")
         x = (i % 4) * 258
         y = (i // 4) * 167
         img.paste(img_i, (x, y))
-        os.remove("Test_Images/" + str(i) + ".png")
+        if reset:
+            os.remove("Test_Images/" + str(i) + ".png")
     for SET in output:
         k = 0
         for i in range(3):
             for j in range(i + 1, 3):
                 k += 1
-                # to prevent overlap jitter the lines
-                x1 = (SET[i] % 4) * 258 + 129 + torch.randint(-5, 5, (1,)).item()
-                x2 = (SET[j] % 4) * 258 + 129 + torch.randint(-5, 5, (1,)).item()
-                y1 = (SET[i] // 4) * 167 + 83 + torch.randint(-5, 5, (1,)).item()
-                y2 = (SET[j] // 4) * 167 + 83 + torch.randint(-5, 5, (1,)).item()
-                if k < 3:
-                    draw = ImageDraw.Draw(img)
-                    draw.line((x1, y1, x2, y2), fill=color_dict[output.index(SET)], width=5)
-    # img.show()
-    img.save(image_name[:image_name.find("_")] + "_solved.png")
+                if k > 2:
+                    continue
+                x1 = (SET[i] % 4) * 258 + 129
+                x2 = (SET[j] % 4) * 258 + 129
+                y1 = (SET[i] // 4) * 167 + 83
+                y2 = (SET[j] // 4) * 167 + 83
+                draw = ImageDraw.Draw(img)
+                draw.line((x1, y1, x2, y2), fill=color_dict[output.index(SET)], width=5)
+    if reset:
+        img.save(image_name[:image_name.find("_")] + "_solved.png")
+    else:
+        img.show()
 
 
 def main(image):
+    if verbose:
+        print("Starting", '-' * 41)
     # check if folder test images is empty
     if len(os.listdir("Test_Images")) == 0:
         img1 = Image.open(image)
@@ -135,8 +142,10 @@ def main(image):
     results = classify()
     output = find(results)
     show(output, image)
+    if verbose:
+        print("Done", '-' * 45)
 
 
 if __name__ == "__main__":
     condition = True
-    main("feb18_game.png")
+    main("feb27_game.png")
